Correct all instances of dataFlags to type unsigned long
Let gradientScore method copy ROI instead of referencing
Add escape in regionProps if no objects found
Add reset member method of ObjectFinder to Analyzer reset method
Add default constructor to tracker

+++ src/CellSorter/lib/mathlab.h
@@ -28,7 +28,7 @@ enum regionPropTypes {
 
 void bwareaopen(cv::Mat& im, double size);
 
-int regionProps(const cv::Mat& img, const int& dataFlags, DataContainer& dc);
+int regionProps(const cv::Mat& img, const unsigned long& dataFlags, DataContainer& dc);
 
 double gradientScore(const cv::Mat& img, const cv::Rect& roi);
 double verticalSymmetry(const cv::Mat& img, const cv::Rect& roi, const double& majorAxisLength);
